refactor: transform vehicle module
